fix: split training into training feature computation and data builder for recommendation engine
